Refactor: Move lineBreak extension function
The `lineBreak` extension function for `TextFieldValue` has been moved from `InputComponents.kt` to a new file `TextFieldValueExt.kt` within the `com.bitchat.android.core.ui.utils` package.

This improves code organization by grouping UI-related utility functions.

Additionally, a Log.d statement for debugging key events in `MessageInput` has been removed.
